Refactore code (CLEAREST CODE)
