--- monster.py
@@ -1,1 +1,0 @@
-# store monster_base here
